requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Database dependencies
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# WebSocket and networking
websockets==11.0.3
aiohttp==3.9.1

# Redis and caching
redis==5.0.1

# Trading and market data
alpaca-py==0.13.3
pandas==2.1.4
numpy==1.26.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and logging
python-json-logger==2.0.7
psutil==5.9.6

# Additional utilities
email-validator==2.1.0